moving responsive Button examples into their own example page
